refactor (ontology cache): move class and interfaces to separate file
