+++ test/data/api/v2/mockOntology.ts
@@ -1,9 +1,9 @@
 import { JsonConvert, OperationMode, ValueCheckingMode } from "json2typescript";
 import { PropertyMatchingRule } from "json2typescript/src/json2typescript/json-convert-enums";
 import {
-    ResourceClassDefinitionWithPropertyDefinition,
-    IResourceClassAndPropertyDefinitions, IHasPropertyWithPropertyDefinition
-} from "../../../../src/cache/OntologyCache";
+    IResourceClassAndPropertyDefinitions,
+    ResourceClassDefinitionWithPropertyDefinition
+} from "../../../../src/cache/ontology-cache/resource-class-definition-with-property-definition";
 import { Constants } from "../../../../src/models/v2/Constants";
 import { IHasProperty } from "../../../../src/models/v2/ontologies/class-definition";
 import { OntologyConversionUtil } from "../../../../src/models/v2/ontologies/OntologyConversionUtil";
